requirement.txt
matplotlib==3.7.5             
numpy==1.24.4     
opencv-python==4.10.0.84  
osrf-pycommon1==0.1.11     
tuya-connector-python==0.1.2      
tuya-iot-py-sdk==0.6.6           
ultralytics==8.3.49     
 

